Creating front page for app
