+++ app/src/main/res/layout/activity_welcome.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     android:id="@+id/activity_welcome"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -8,40 +9,14 @@
     android:orientation="vertical"
     tools:context=".WelcomeActivity">
 
-    <TextView
-        android:id="@+id/welcomeText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:text="@string/welcome"
-        android:textSize="45sp"
-        android:textAlignment="center"
-        android:layout_marginTop="150dp"
-        android:layout_marginBottom="50dp"
-        android:textColor="@color/charcoal"
-        />
 
-    <Button
-        style="@style/WelcomeButtons"
-        android:id="@+id/btn_login"
-        android:layout_below="@id/welcomeText"
-        android:text="@string/login"
-        />
-
-    <Button
-        style="@style/WelcomeButtons"
-        android:id="@+id/btn_signup"
-        android:layout_below="@id/btn_login"
-        android:text="@string/signup"
-        android:backgroundTint="@color/seafoam"
-        />
-
-    <Button
-        style="@style/WelcomeButtons"
-        android:id="@+id/btn_guest"
-        android:layout_below="@+id/btn_signup"
-        android:text="@string/use_as_guest"
-        android:backgroundTint="@color/seafoam"
-        />
+    <FrameLayout
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </RelativeLayout>
