+++ app/src/main/res/values/strings.xml
@@ -38,6 +38,8 @@
     <string name="next_step">Next Step</string>
     <string name="diagnosis_title">A possible Diagnosis is...</string>
     <string name="year">Year:</string>
+    <string name="make">Make:</string>
+    <string name="model">Model:</string>
     <string name="title_activity_help_map">Map</string>
     <string name="back_to_home_text">Back home</string>
     <string name="mechanics">Mechanics</string>
